Refactor: styles according of its sections and create new files to each section component
