+++ src/components/fedmg/components/ControlSection.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+
+export const ControlSection = ({ response, handler: generatePass }) => {
+	return (
+		<div className='button-container'>
+			<div className='message-container'>
+				<span className='response-message'>{response}</span>
+			</div>
+			<button className='generate-button' onClick={generatePass}>
+				Generate
+			</button>
+		</div>
+	)
+}
