+++ src/components/fedmg/response.js
@@ -1,0 +1,5 @@
+export const responses = {
+	password: '',
+	status: false,
+	message: ''
+}
