+++ src/components/fedmg/styles/input.css
@@ -1,0 +1,28 @@
+.fedmg-main .input-container {
+	display: grid;
+	gap: clamp(0.82em, 3vw, 1.1em) 0;
+}
+
+.label-container {
+	display: grid;
+	grid-template-columns: 1fr 1fr;
+}
+
+.fedmg-main .range-input {
+	user-select: none;
+}
+
+.fedmg-main .range-lbl {
+	text-align: start;
+
+	/* clamp(13.5px, 3vw, 19px) */
+	font-size: clamp(1.05em, 3vw, 1.16em);
+}
+
+.fedmg-main .range-span {
+	font-size: clamp(0.96em, 3vw, 1.1em);
+	border: 1px solid #fff;
+	background: rgb(0 0 0 / 30%);
+	border-radius: 0.29em;
+	font-family: 'Courier new', Courier, sans-serif;
+}
