+++ src/components/fedmg/styles/message.css
@@ -1,0 +1,7 @@
+/* 
+.fedmg-main .response-message {
+}
+.fedmg-main .response-message_success {
+}
+.fedmg-main .response-message_wrong {
+} */
