+++ src/components/fedmg/styles/password.css
@@ -1,0 +1,41 @@
+/* .fedmg-main .container-pass, .fedmg-main .clipboard-button, .fedmg-main  {
+	
+} */
+
+.fedmg-main .password-container {
+	margin: clamp(0.82em, 3vw, 1.1em) 0;
+	background-color: rgb(0 0 0 / 30%);
+	display: grid;
+	grid-template-columns: 1fr 1fr 1fr;
+	gap: 0.9em;
+}
+
+.fedmg-main .container-pass {
+	grid-column: 1/3;
+	border: 1px solid #fff;
+	padding: clamp(0.1em, 2vw, 0.3em) 0;
+}
+
+.fedmg-main .password-response {
+	font-size: clamp(0.96em, 3vw, 1.1em);
+	font-family: 'Courier new', Courier, sans-serif;
+	text-overflow: hidden;
+}
+
+.fedmg-main .clipboard-button {
+	grid-column: 3;
+	border: 1px solid #fff;
+}
+
+/* status */
+.fedmg-main .clipboard-button:hover {
+	background-color: rgb(255 255 255 / 30%);
+	border: 1px solid rgb(255 255 255 / 70%);
+	color: rgb(0 0 0);
+}
+
+.fedmg-main .clipboard-button:active {
+	background-color: rgb(0 0 0 / 30%);
+	border: 1px solid #fff;
+	color: #fff;
+}
